requirements.txt
astroid==2.14.2
click==8.1.7
colorama==0.4.6
dill==0.3.7
Flask==2.2.2
Flask-Assets==2.1.0
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
isort==5.9.3
itsdangerous==2.0.1
jinja-partials==0.1.1
Jinja2==3.1.2
lazy-object-proxy==1.6.0
MarkupSafe==2.1.1
mccabe==0.7.0
pip==23.3
platformdirs==3.10.0
psycopg2==2.9.3
pylint==2.16.2
pytailwindcss==0.2.0
python-dotenv==0.21.0
setuptools==68.0.0
SQLAlchemy==2.0.21
tomlkit==0.11.1
typing_extensions==4.7.1
webassets==2.0
Werkzeug==2.2.3
wheel==0.41.2
wrapt==1.14.1
